+++ Code/Project/Project/AddParticipantsPage.xaml.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.VisualBasic;
+using Project.Interfaces;
+using Project.Other;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
