+++ Code/Project/Project/FileWritePage.xaml.cs
@@ -15,6 +15,9 @@ using System.Windows.Shapes;
 using System.Windows.Forms;
 using System.IO;
 using Newtonsoft.Json;
+using Project.Other;
+using Project.Interfaces;
+using Task = Project.Other.Task;
 
 namespace Project
 {
