--- Code/Project/Project/IFileWritePage.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Project
-{
-    public interface IFileWritePage
-    {
-        public void NewFileWritePage(List<string> employees);
-    }
-}
