+++ Code/Project/Project/Interfaces/IAddParticipantsPage.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Project
+namespace Project.Interfaces
 {
     public interface IAddParticipantsPage
     {
