+++ Code/Project/Project/MainWindow.xaml.cs
@@ -12,6 +12,8 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Project.Interfaces;
+using Project.Other;
 
 namespace Project
 {
