+++ Code/Project/Project/NewTaskPage.xaml.cs
@@ -13,6 +13,8 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Project.Interfaces;
+using Project.Other;
 
 namespace Project
 {
