+++ Code/Project/Project/Other/Global.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Navigation;
 
-namespace Project
+namespace Project.Other
 {
     public static class Global
     {
