+++ Code/Project/Project/Other/Task.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Project.Other
+{
+    public class Task
+    {
+        public string Name = "";
+        public string Description = "";
+        public int Rating;
+    }
+}
