+++ Code/Project/Project/PlanningPokerPage.xaml.cs
@@ -14,6 +14,9 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Project.Interfaces;
+using Project.Other;
+using Task = Project.Other.Task;
 
 namespace Project
 {
